% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\name{lists}
\alias{lists}
\alias{dnd_chars_unique_list}
\alias{dnd_chars_all_list}
\alias{dnd_chars_singleclass_list}
\alias{dnd_chars_multiclass_list}
\title{D&D character lists}
\description{
\code{dnd_chars_all_list}, \code{dnd_chars_unique_list},
\code{dnd_chars_singleclass_list} and \code{dnd_chars_multiclass_list} \code{list}s
 contains character data submitted to \url{https://oganm.com/shiny/printSheetApp} and
\url{https://oganm.com/shiny/interactiveSheet}.

These lists include the same data that is included in the \code{data.frame}
objects in this package with complex data fields such as spells are split into their
components for easier access.

\itemize{
    \item{dnd_chars_all_list: All characters. May contain duplicates}
    \item{dnd_chars_unique_list: Filtered version of \code{dnd_chars_all} to remove possible duplicates. Highest level variant of the characters are kept}
    \item{dnd_chars_singleclass_list: Subset of \code{dnd_chars_unique} that only includes single classed characters}
    \item{dnd_chars_multiclass_list: Subset of \code{dnd_chars_unique} that only includes multiclassed characters}
}
}
\seealso{
\code{\link{dnd_chars_unique)}}, \code{\link{dnd_chars_all}}, \code{\link{dnd_chars_singleclass}}, \code{\link{dnd_chars_multiclass}}
}
